Kotlin upgrade

+++ app/src/main/java/com/quigglesproductions/secureimageviewer/models/enhanced/file/EnhancedDatabaseFile.java
@@ -1,10 +1,8 @@
 package com.quigglesproductions.secureimageviewer.models.enhanced.file;
 
-import android.os.Parcel;
-
 import androidx.annotation.VisibleForTesting;
 
-import com.quigglesproductions.secureimageviewer.models.enhanced.datasource.LocalFileDataSource;
+import com.quigglesproductions.secureimageviewer.datasource.file.LocalFileDataSource;
 
 import java.io.File;
 import java.time.LocalDateTime;
@@ -38,8 +36,8 @@ public class EnhancedDatabaseFile extends EnhancedFile implements IDatabaseFile{
         setDataSource(new LocalFileDataSource(this));
     }
     @Override
-    public long getId() {
-        return id;
+    public Long getId() {
+        return (long) id;
     }
     @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
     public void setId(int id){
