+++ app/src/main/java/com/quigglesproductions/secureimageviewer/models/enhanced/file/EnhancedFile.java
@@ -1,15 +1,13 @@
 package com.quigglesproductions.secureimageviewer.models.enhanced.file;
 
-import android.os.Parcel;
-
 import androidx.annotation.NonNull;
 
 import com.google.gson.annotations.SerializedName;
+import com.quigglesproductions.secureimageviewer.datasource.file.IFileDataSource;
+import com.quigglesproductions.secureimageviewer.models.ItemBaseModel;
 import com.quigglesproductions.secureimageviewer.models.enhanced.EnhancedArtist;
 import com.quigglesproductions.secureimageviewer.models.enhanced.EnhancedCategory;
 import com.quigglesproductions.secureimageviewer.models.enhanced.EnhancedSubject;
-import com.quigglesproductions.secureimageviewer.models.enhanced.datasource.IFileDataSource;
-import com.quigglesproductions.secureimageviewer.models.ItemBaseModel;
 import com.quigglesproductions.secureimageviewer.models.enhanced.metadata.FileMetadata;
 
 import java.time.LocalDateTime;
@@ -155,8 +153,8 @@ public class EnhancedFile implements ItemBaseModel,IDisplayFile {
     }
 
     @Override
-    public long getId() {
-        return onlineId;
+    public Long getId() {
+        return (long) onlineId;
     }
 
     public IFileDataSource getDataSource() {
